requirements.txt
streamlit>=1.36.0
pandas>=2.2.0
numpy>=1.26.0
regex>=2023.12.25
openai>=1.40.0
pillow>=10.3.0
easyocr>=1.7.1
pdf2image>=1.17.0
pytesseract>=0.3.10
PyPDF2>=3.0.1
email-validator>=2.1.0
# Gmail API is optional - using IMAP instead (free, no credits)
# google-api-python-client>=2.100.0
# google-auth-httplib2>=0.1.1
# google-auth-oauthlib>=1.1.0
# PaddleOCR is heavy and problematic on macOS; enable on linux/win x86_64
paddlepaddle==2.6.1; platform_system != 'Darwin' and platform_machine == 'x86_64'
paddleocr>=2.7.0; platform_system != 'Darwin' and platform_machine == 'x86_64'
